refactor(review): extract recommendation engine modules
